pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "meridianiq"
version = "0.9.0"
description = "The intelligence standard for project schedules — open-source schedule intelligence from validation to prediction"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "Vitor Maia Rodovalho", email = "vitor.rodovalho@outlook.com"},
]
keywords = ["primavera", "p6", "xer", "schedule", "cpm", "construction"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "pydantic>=2.0",
    "networkx>=3.0",
    "numpy>=1.24",
    "supabase>=2.0",
    "python-dotenv>=1.0",
    "python-multipart",
    "sentry-sdk[fastapi]>=2.0",
    "PyJWT[crypto]>=2.0",
]

[project.optional-dependencies]
api = ["fastapi>=0.100", "uvicorn[standard]"]
export = ["pandas>=2.0"]
reports = ["weasyprint>=60.0"]
dev = ["pytest>=8.0", "pytest-cov", "ruff", "mypy", "httpx>=0.24"]
all = ["meridianiq[api,export,reports,dev]"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
